requirements.txt
aider-chat==0.16.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
autograd==1.8.0
autoray==0.6.11
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
ConfigArgParse==1.7.5
cryptography @ file:///home/builder/.termux-build/python-cryptography/src
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
fastapi==0.135.1
Flask==3.1.3
flask-cors==6.0.2
frozenlist==1.8.0
gallery_dl==1.31.9
gast==0.5.4
gitdb==4.0.12
GitPython==3.1.46
grep-ast==0.9.0
h11==0.16.0
hatch-requirements-txt==0.4.1
hatchling==1.29.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
internetarchive==5.8.0
itsdangerous==2.2.0
jaraco.context==6.1.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml @ file:///home/builder/.termux-build/python-lxml/src
markdown-it-py==4.0.0
MarkupSafe==3.0.3
maturin==1.12.6
mdurl==0.1.2
meson==1.10.2
meson-python==0.19.0
more-itertools==10.8.0
motor==3.7.1
multidict==6.7.1
networkx==3.6.1
ninja==1.13.0
nltk==3.9.3
numpy==2.4.3
oauthlib==3.3.1
openai==0.27.6
openpyxl==3.1.5
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
PennyLane==0.36.0
PennyLane-Lightning==0.36.0
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pymongo==4.16.0
PyPDF2==3.0.1
pyproject-metadata==0.11.0
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rustworkx==0.17.1
scipy==1.17.1
semantic-version==2.10.0
setuptools==82.0.1
setuptools-rust==1.12.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
termcolor==3.3.0
tiktoken==0.12.0
toml==0.10.2
tomlkit==0.14.0
tqdm==4.67.3
trove-classifiers==2026.1.14.14
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
wolframalpha==5.1.3
wolframclient==1.4.0
xlsxwriter==3.2.9
xmltodict==1.0.4
yarl==1.23.0
